get_example_response: move this util to toplevel for reuse
