+++ ckanfunctionaltests/api/test_schemas.py
@@ -1,23 +1,10 @@
-from copy import deepcopy
-from functools import lru_cache
 import json
 import os.path
 
 import jsonschema
 import pytest
 
-from ckanfunctionaltests.api import validate_against_schema
-
-
-@lru_cache()
-def _get_example_response_inner(filename: str):
-    with open(os.path.join(os.path.dirname(__file__), "example_responses", filename), "rb") as f:
-        return json.load(f)
-
-
-def _get_example_response(filename: str):
-    # returning a deepcopy allows the caller to mutate the response safely
-    return deepcopy(_get_example_response_inner(filename))
+from ckanfunctionaltests.api import validate_against_schema, get_example_response
 
 
 @pytest.mark.parametrize("response_filename,schema_name", (
@@ -34,7 +21,7 @@ def _get_example_response(filename: str):
     ("i18n.json", "i18n",),
 ))
 def test_responses_pass_respective_schemas(response_filename, schema_name):
-    validate_against_schema(_get_example_response(response_filename), schema_name)
+    validate_against_schema(get_example_response(response_filename), schema_name)
 
 
 @pytest.mark.parametrize("nonslug_value", (
@@ -42,7 +29,7 @@ def test_responses_pass_respective_schemas(response_filename, schema_name):
     "One fine morning in the month of May",
 ))
 def test_package_list_must_be_slugs(nonslug_value):
-    example_response = _get_example_response("package_list.json")
+    example_response = get_example_response("package_list.json")
     example_response["result"][2] = nonslug_value
 
     with pytest.raises(jsonschema.ValidationError):
@@ -50,7 +37,7 @@ def test_package_list_must_be_slugs(nonslug_value):
 
 
 def test_package_show_org_must_follow_org_schema():
-    example_response = _get_example_response("package_show.json")
+    example_response = get_example_response("package_show.json")
 
     # note non-existent leap-day
     example_response["result"]["organization"]["created"] = "2019-02-29T16:00:00.123Z"
@@ -60,7 +47,7 @@ def test_package_show_org_must_follow_org_schema():
 
 
 def test_organization_show_is_organization():
-    example_response = _get_example_response("organization_show.json")
+    example_response = get_example_response("organization_show.json")
 
     example_response["result"]["is_organization"] = False
 
@@ -73,7 +60,7 @@ def test_organization_show_is_organization():
     "package_show",
 ))
 def test_missing_ids(endpoint):
-    example_response = _get_example_response(f"{endpoint}.json")
+    example_response = get_example_response(f"{endpoint}.json")
 
     del example_response["result"]["id"]
 
@@ -86,7 +73,7 @@ def test_missing_ids(endpoint):
     "package_show",
 ))
 def test_unknown_state(endpoint):
-    example_response = _get_example_response(f"{endpoint}.json")
+    example_response = get_example_response(f"{endpoint}.json")
 
     example_response["result"]["state"] = "lockdown"
 
@@ -102,7 +89,7 @@ def test_unknown_state(endpoint):
     "package_search",
 ))
 def test_missing_success(endpoint):
-    example_response = _get_example_response(f"{endpoint}.json")
+    example_response = get_example_response(f"{endpoint}.json")
 
     del example_response["success"]
 
@@ -111,7 +98,7 @@ def test_missing_success(endpoint):
 
 
 def test_package_search_noninteger_facet():
-    example_response = _get_example_response("package_search.json")
+    example_response = get_example_response("package_search.json")
 
     example_response["result"]["facets"]["organization"]["natural-england"] = "2"
 
@@ -120,7 +107,7 @@ def test_package_search_noninteger_facet():
 
 
 def test_package_search_empty_tag():
-    example_response = _get_example_response("package_search.json")
+    example_response = get_example_response("package_search.json")
 
     example_response["result"]["results"][2]["tags"].append({})
 
@@ -129,7 +116,7 @@ def test_package_search_empty_tag():
 
 
 def test_format_autocomplete_result_missing_format():
-    example_response = _get_example_response("format_autocomplete.json")
+    example_response = get_example_response("format_autocomplete.json")
 
     del example_response["ResultSet"]["Result"][1]["Format"]
